fix img timestamp

+++ include/mynteye/types.h
@@ -410,7 +410,7 @@ struct MYNTEYE_API ImgData {
   /** Image frame id */
   std::uint16_t frame_id;
   /** Image timestamp in 0.01ms */
-  std::uint32_t timestamp;
+  std::uint64_t timestamp;
   /** Image exposure time, virtual value in [1, 480] */
   std::uint16_t exposure_time;
